requirements.txt
Django==3.0.4
django_celery_beat==2.1.0
djangorestframework==3.13.1
feedparser==6.0.8
python_dateutil==2.8.2
psycopg2==2.8.6
celery==4.4.2
redis==3.4.1